model: Move generated SPDX files also to the "spdx" directory
