+++ model/build.gradle
@@ -64,10 +64,10 @@ task generateSpdxLicenseList {
 
         println "Found ${spdxLicenseIds.size()} SPDX license identifiers."
 
-        def licenseEnumFile = file('src/main/kotlin/config/SpdxLicense.kt')
+        def licenseEnumFile = file('src/main/kotlin/spdx/SpdxLicense.kt')
         licenseEnumFile.write(getLicenseHeader())
         licenseEnumFile.append("""\
-            package com.here.ort.model.config
+            package com.here.ort.model.spdx
 
             /**
              * An enum containing all SPDX license IDs. This class is generated by the Gradle task
@@ -106,10 +106,10 @@ task generateSpdxLicenseExceptionList {
 
         println "Found ${spdxLicenseExceptionIds.size()} SPDX license exception identifiers."
 
-        def licenseExceptionEnumFile = file('src/main/kotlin/config/SpdxLicenseException.kt')
+        def licenseExceptionEnumFile = file('src/main/kotlin/spdx/SpdxLicenseException.kt')
         licenseExceptionEnumFile.write(getLicenseHeader())
         licenseExceptionEnumFile.append("""\
-            package com.here.ort.model.config
+            package com.here.ort.model.spdx
 
             /**
              * An enum containing all SPDX license exception IDs. This class is generated by the Gradle task
