+++ model/src/main/kotlin/spdx/SpdxLicense.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.model.config
+package com.here.ort.model.spdx
 
 /**
  * An enum containing all SPDX license IDs. This class is generated by the Gradle task
