+++ model/src/main/kotlin/spdx/SpdxLicenseException.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.model.config
+package com.here.ort.model.spdx
 
 /**
  * An enum containing all SPDX license exception IDs. This class is generated by the Gradle task
